+++ packages/components/src/components/navigation/link-pure/link-pure-styles.ts
@@ -1,8 +1,51 @@
+import type { Styles } from 'jss';
 import type { BreakpointCustomizable } from '../../../utils';
 import type { AlignLabel, LinkButtonPureIconName, TextSize, ThemeExtendedElectricDark } from '../../../types';
 import { buildSlottedStyles, getCss, mergeDeep } from '../../../utils';
-import { getFocusSlottedPseudoStyles } from '../../../styles';
+import { getInset, getThemedColors, getTransition } from '../../../styles';
 import { getLinkButtonPureStyles } from '../../../styles/link-button-pure-styles';
+
+export type GetFocusSlottedPseudoStylesOptions = {
+  color?: string;
+  offset?: number;
+};
+
+/**
+ * this hack is only needed for Safari which does not support pseudo elements in slotted context (https://bugs.webkit.org/show_bug.cgi?id=178237) :-(
+ */
+export const getFocusSlottedPseudoStyles = (opts?: GetFocusSlottedPseudoStylesOptions): Styles<'& a'> => {
+  const { offset: outlineOffset, color: outlineColor }: GetFocusSlottedPseudoStylesOptions = {
+    color: 'currentColor',
+    offset: 2,
+    ...opts,
+  };
+
+  return {
+    '& a': {
+      display: 'block',
+      position: 'static',
+      textDecoration: 'none',
+      font: 'inherit',
+      // color: 'inherit', // TODO: chrome hover bug. Use when fixed.
+      transition: getTransition('color'), // TODO: chrome hover bug. Remove when fixed.
+      outline: 'transparent none',
+      '&::before': {
+        content: '""',
+        display: 'block',
+        position: 'absolute',
+        ...getInset(),
+        outline: '1px solid transparent',
+        outlineOffset: `${outlineOffset}px`,
+      },
+      '&:focus::before': {
+        outlineColor,
+      },
+      '&:focus:not(:focus-visible)::before': {
+        outlineColor: 'transparent',
+      },
+    },
+  };
+};
 
 export const getComponentCss = (
   icon: LinkButtonPureIconName,
@@ -15,6 +58,8 @@ export const getComponentCss = (
   hasSlottedAnchor: boolean,
   theme: ThemeExtendedElectricDark
 ): string => {
+  const { baseColor, hoverColor, activeColor } = getThemedColors(theme);
+
   return getCss(
     mergeDeep(
       getLinkButtonPureStyles(
@@ -32,8 +77,16 @@ export const getComponentCss = (
       // TODO:V3 should be removed, we shouldn't support this although some CMS are rendering an <a> with a wrapped <p>. Instead CMS output shall be post processed because it's necessary to use the PDS component anyway.
       {
         '@global': {
-          '::slotted(p)': {
-            margin: 0,
+          '::slotted': {
+            '&(p)': {
+              margin: 0,
+            },
+            '&(a)': {
+              color: active ? activeColor : baseColor, // TODO: chrome hover bug. Remove when fixed.
+            },
+            '&(a:hover)': {
+              color: hoverColor, // TODO: chrome hover bug. Remove when fixed.
+            },
           },
         },
       }