+++ packages/components/CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [2.9.2-rc.1] - 2022-03-23
 
+#### Fixed
+- Bug caused by Chrome where hover styles of `Link Pure` are not displayed correctly
+
 ### [2.9.2-rc.0] - 2022-03-22
 
 #### Added